+++ CHANGES.md
@@ -1,5 +1,7 @@
-- Move sort option in recipe fiter to the bottom
+- Add option to track recipes per quality
 - Add restock settings to item and enchant recipe forms
 - Add filter option to see all restocking recipes and to mark all as tracked
 - Add filter option to see all tracked recipes and to craft next batch
+- Move sort option in recipe fiter to the bottom
+- Show recipe qualities separately when sorting recipes by optimization goal
 - Fix showing qualities in recipe tracker for recipes that don't have any